improve some type information to reduce backedges; fixes #29166
reuse `Base._array_for` in lowering instead of manually inlining it

[ci skip]

+++ base/strings/io.jl
@@ -113,7 +113,7 @@ function print_to_string(xs...)
     if isempty(xs)
         return ""
     end
-    siz = 0
+    siz::Int = 0
     for x in xs
         siz += tostr_sizehint(x)
     end
